requirements.txt
Django==1.10.2
django-ajax-selects==1.5.0
django-related-admin==0.3.1
Pillow==3.4.1
pkg-resources==0.0.0
six==1.10.0
